enable switches by default
Added a assumed state to mqtt subscriptions. When mqtt does not return a
value initially we will assume this value. The service will also push
the mqtt value to the topic so everybody is in sync.

Also added a HassStatus service to make the automations reload as soon
as detected a home assistant restart.